Create Merger and SingleQubitGatesMerger classes, move merge functionality to said classes
