+++ opensquirrel/passes/merger/__init__.py
@@ -1,0 +1,7 @@
+from opensquirrel.passes.merger.general_merger import Merger
+from opensquirrel.passes.merger.single_qubit_gates_merger import SingleQubitGatesMerger
+
+__all__ = [
+    "Merger",
+    "SingleQubitGatesMerger",
+]
